fixing

+++ src/main/java/lb/project/lb6_server/server/data/WorkersRepository.java
@@ -1,7 +1,9 @@
 package lb.project.lb6_server.server.data;
 
 import jakarta.annotation.PostConstruct;
+import jakarta.transaction.Transactional;
 import lb.project.lb6_server.lib.entities.Person;
+import lb.project.lb6_server.lib.entities.User;
 import lb.project.lb6_server.lib.entities.Worker;
 import lb.project.lb6_server.server.data.savers.db.WorkerJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +11,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
+@Transactional
 public class WorkersRepository implements IWorkersRepository{
 
     @Autowired
@@ -26,32 +29,38 @@ public class WorkersRepository implements IWorkersRepository{
     }
 
     @Override
-    public void insert(int key, Worker worker) {
-        hashtable.insert(key, worker);
+    public boolean insert(int key, Worker worker, User user) {
+        if(!hashtable.insert(key, worker, user))
+            return false;
+
         jpaRepository.saveAndFlush(worker);
+        return true;
     }
 
     @Override
-    public void update(int key, Worker worker) {
+    public boolean update(int key, Worker worker, User user) {
 
-        if(!worker.getOwner().equals(hashtable.getByKey(key).getOwner()))
-            return;
+        if(!hashtable.update(key, worker, user))
+            return false;
 
-        hashtable.update(key, worker);
         jpaRepository.saveAndFlush(hashtable.getByKey(key));//check
+        return true;
     }
 
     @Override
-    public void remove(int key) {
-        hashtable.remove(key);
+    public boolean remove(int key, User user) {
+        if (!hashtable.remove(key, user))
+            return false;
+
         jpaRepository.delete(hashtable.getByKey(key));
         jpaRepository.flush();
+        return true;
     }
 
     @Override
-    public void clear() {
-        hashtable.clear();
-        jpaRepository.deleteAll();
+    public void clear(User user) {
+        hashtable.clear(user);
+        jpaRepository.deleteAllByOwner(user);
         jpaRepository.flush();
     }
 
@@ -61,28 +70,32 @@ public class WorkersRepository implements IWorkersRepository{
     }
 
     @Override
-    public void removeLowerWorkers(Worker worker) {
-        List<Worker> workersToRemove = hashtable.getLowerWorkers(worker);
-        hashtable.removeLowerWorkers(worker);
+    public void removeLowerWorkers(Worker worker, User user) {
+        List<Worker> workersToRemove = hashtable.getLowerWorkersByUser(worker, user);
+        hashtable.removeLowerWorkers(worker, user);
         jpaRepository.deleteAll(workersToRemove);
         jpaRepository.flush();
     }
 
     @Override
-    public void replaceWithGreaterWorker(int key, Worker worker) {
+    public boolean replaceWithGreaterWorker(int key, Worker worker, User user) {
         Worker currentWorker = hashtable.getByKey(key);
         if (currentWorker.compareTo(worker) > 0) {
-            hashtable.insert(key, worker);
+            if(!hashtable.insert(key, worker, user))
+                return false;
+
             jpaRepository.delete(currentWorker);
             jpaRepository.saveAndFlush(worker);
+
+            return true;
         }
+        return false;
     }
 
     @Override
-    public void removeWorkersWithGreaterKey(int key) {
-        List<Worker> workers = hashtable.getListOfWorkersWithGreaterKey(key);
-        hashtable.removeWorkersWithGreaterKey(key);
-        jpaRepository.deleteAll(workers);
+    public void removeWorkersWithGreaterKey(int key, User user) {
+        hashtable.removeWorkersWithGreaterKey(key, user);
+        jpaRepository.deleteAllByHashtableKeyGreaterThan(key);
         jpaRepository.flush();
 
     }
